+++ server/component/style/StyleModel.php
@@ -113,6 +113,7 @@ class StyleModel extends BaseModel implements IStyleModel
         $this->params = $params;
         $this->id_page = $id_page; 
         $this->entry_record = $entry_record;
+<<<<<<< HEAD
         $this->style_name = $this->get_style_name_by_section_id($id);   
         if (DEBUG) {
             PerformanceLogger::startTimer(
@@ -125,9 +126,24 @@ class StyleModel extends BaseModel implements IStyleModel
         }     
         if(isset($params['parent_id'])){
             $this->parent_id = $params['parent_id'];
+=======
+        $this->initialize();
+                
+    }
+
+    /* Private Methods ********************************************************/    
+
+    public function initialize()
+    {
+        $class = get_class($this);
+        $this->services->get_clockwork()->startEvent("[$class][initialize]");
+        $this->style_name = $this->get_style_name_by_section_id($this->section_id);        
+        if(isset($this->params['parent_id'])){
+            $this->parent_id = $this->params['parent_id'];
+>>>>>>> main
         }
-        if(isset($params['relation'])){
-            $this->relation = $params['relation'];
+        if(isset($this->params['relation'])){
+            $this->relation = $this->params['relation'];
         }
         if(!$this->is_cms_page())
         {
@@ -135,10 +151,10 @@ class StyleModel extends BaseModel implements IStyleModel
             $_SESSION['language'] = $_SESSION['user_language'];
         }
         $this->db_fields['id'] = array(
-            "content" => $id,
+            "content" => $this->section_id,
             "type" => "internal",
         );                                
-        $fields = $this->db->fetch_section_fields($id);
+        $fields = $this->db->fetch_section_fields($this->section_id);
         $this->set_db_fields($fields);   
         if(!$this->style_name){
             return;
@@ -149,6 +165,7 @@ class StyleModel extends BaseModel implements IStyleModel
         if (($this->is_cms_page() || $this->condition_result['result'])) {
             $this->loadChildren($this->entry_record);
         }
+<<<<<<< HEAD
         if (DEBUG && $this->style_name) {
             PerformanceLogger::endTimer(
                 'style-model',
@@ -163,6 +180,11 @@ class StyleModel extends BaseModel implements IStyleModel
     }
 
     /* Private Methods ********************************************************/    
+=======
+        $class = get_class($this);
+        $this->services->get_clockwork()->endEvent("[$class][initialize]");
+    }   
+>>>>>>> main
 
     /**
      * Get entries values if there are any set